requirements.txt
streamlit>=1.32.0
openai>=1.98.0
pinecone>=3.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
numpy>=1.24.3
beautifulsoup4>=4.12.0
requests>=2.31.0
Pillow>=10.0.0
easyocr>=1.7.0
openai-whisper>=20250625
sounddevice>=0.4.6 